stories_microservices/stories_microservices_post_service/requirements.txt
alembic==1.4.3
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
six==1.15.0
SQLAlchemy==1.3.22
text-unidecode==1.3
Werkzeug==1.0.1
